Tests/LibWeb/Text/expected/css/custom-property-cascade-order.txt

normal inline: 3px
normal source order: 5px
important inline: rgb(255, 0, 255)
important layer order: rgb(0, 128, 0)
revert-layer: 10px
normal outer context: rgb(0, 128, 0)
important inner context: rgb(0, 0, 255)
